+++ app/src/main/java/com/kevinj1008/fitnessch/activities/FitnesschLoginActivity.java
@@ -2,10 +2,7 @@ package com.kevinj1008.fitnessch.activities;
 
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.content.IntentFilter;
-import android.graphics.Color;
 import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.support.annotation.NonNull;
@@ -15,8 +12,6 @@ import android.support.v7.app.AlertDialog;
 import android.util.Log;
 
 import android.view.View;
-import android.view.Window;
-import android.view.WindowManager;
 
 import android.widget.Toast;
 
@@ -46,7 +41,6 @@ import com.kevinj1008.fitnessch.R;
 import com.kevinj1008.fitnessch.UserExistCallback;
 import com.kevinj1008.fitnessch.util.Constants;
 import com.kevinj1008.fitnessch.util.ForceUpdateChecker;
-import com.kevinj1008.fitnessch.util.NetworkChangeReceiver;
 import com.kevinj1008.fitnessch.util.NetworkUtils;
 import com.kevinj1008.fitnessch.util.SharedPreferencesManager;
 import java.util.HashMap;
@@ -103,7 +97,7 @@ public class FitnesschLoginActivity extends BaseActivity implements GoogleApiCli
                     }
                 }, 1000);
             } else {
-                Toast.makeText(FitnesschLoginActivity.this, "登入失敗，請檢查網路連線。", Toast.LENGTH_SHORT).show();
+                Toast.makeText(FitnesschLoginActivity.this, R.string.login_network_fail_toast, Toast.LENGTH_SHORT).show();
             }
         } else {
             mGoogleLogInBtn.setVisibility(View.VISIBLE);
@@ -133,7 +127,7 @@ public class FitnesschLoginActivity extends BaseActivity implements GoogleApiCli
                 if (mNetworkUtils.isNetworkAvailable()) {
                     googleSignIn();
                 } else {
-                    Toast.makeText(FitnesschLoginActivity.this, "登入失敗，請檢查網路連線。", Toast.LENGTH_SHORT).show();
+                    Toast.makeText(FitnesschLoginActivity.this, R.string.login_network_fail_toast, Toast.LENGTH_SHORT).show();
                 }
             }
         }
@@ -172,9 +166,9 @@ public class FitnesschLoginActivity extends BaseActivity implements GoogleApiCli
                     user.put("id_token", mGoogleIdToken);
                     user.put("joined_time", FieldValue.serverTimestamp());
                     user.put("db_uid", uid);
-                    user.put("height", "0");
-                    user.put("weight", "0");
-                    user.put("info", "在這裡輸入個人資訊。");
+                    user.put("height", getString(R.string.default_personal_height));
+                    user.put("weight", getString(R.string.default_personal_weight));
+                    user.put("info", getString(R.string.default_personal_info));
 
                     db.collection("users").document(uid).set(user).addOnSuccessListener(new OnSuccessListener<Void>() {
                         @Override
@@ -246,7 +240,7 @@ public class FitnesschLoginActivity extends BaseActivity implements GoogleApiCli
                 firebaseAuthWithGoogle(credential);
             } else {
                 Log.d(Constants.TAG, "Login Unsuccessful. ");
-                Toast.makeText(this, "登入失敗。", Toast.LENGTH_SHORT).show();
+                Toast.makeText(this, R.string.login_fail_toast, Toast.LENGTH_SHORT).show();
             }
         }
     }
@@ -261,10 +255,10 @@ public class FitnesschLoginActivity extends BaseActivity implements GoogleApiCli
                         if (!task.isSuccessful()) {
                             Log.d(Constants.TAG, "Sign in with credential " + task.getException().getMessage());
                             task.getException().printStackTrace();
-                            Toast.makeText(FitnesschLoginActivity.this, "認證失敗", Toast.LENGTH_SHORT).show();
+                            Toast.makeText(FitnesschLoginActivity.this, R.string.login_auth_fail_toast, Toast.LENGTH_SHORT).show();
                         } else {
                             createUserInFirestore();
-                            Toast.makeText(FitnesschLoginActivity.this, "登入成功。", Toast.LENGTH_SHORT).show();
+                            Toast.makeText(FitnesschLoginActivity.this, R.string.login_success_toast, Toast.LENGTH_SHORT).show();
                             Intent intent = new Intent(FitnesschLoginActivity.this, FitnesschActivity.class);
                             intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
                             startActivity(intent);
@@ -312,16 +306,16 @@ public class FitnesschLoginActivity extends BaseActivity implements GoogleApiCli
     @Override
     public void onUpdateNeeded(String updateUrl) {
                 AlertDialog dialog = new AlertDialog.Builder(this)
-                .setTitle("有新版本！")
-                .setMessage("請到 Google Play 下載最新版本，以利最佳體驗。")
-                .setPositiveButton("更新",
+                .setTitle(R.string.title_force_update_dialog)
+                .setMessage(R.string.message_force_update_dialog)
+                .setPositiveButton(R.string.positive_button_force_update_dialog,
                         new DialogInterface.OnClickListener() {
                             @Override
                             public void onClick(DialogInterface dialog, int which) {
                                 redirectStore(Constants.FITNESSCH_STORE_URL);
                             }
                         })
-                .setNegativeButton("晚點",
+                .setNegativeButton(R.string.negative_button_force_update_dialog,
                         new DialogInterface.OnClickListener() {
                             @Override
                             public void onClick(DialogInterface dialog, int which) {
